+++ src/api/bridge/rnastructure.h
@@ -35,7 +35,7 @@ class RNAstructure : public RnaPackage {
   [[nodiscard]] FoldResult Fold(const Primary& r, std::optional<MfeAlg> alg, erg::EnergyCfg cfg,
       const erg::PseudofreeCfg& pf, const trace::TraceCfg& trace_cfg) const override;
 
-  [[nodiscard]] int Subopt(const Primary& r, std::optional<MfeAlg> mfe_alg,
+  [[nodiscard]] int64_t Subopt(const Primary& r, std::optional<MfeAlg> mfe_alg,
       std::optional<SuboptAlg> alg, erg::EnergyCfg cfg, const erg::PseudofreeCfg& pf,
       const subopt::SuboptCallback& fn, subopt::SuboptCfg subopt_cfg) const override;
